.history/requirements_20250226005832.txt
litellm[proxy]
ollama
pyautogen[retrievechat]
tiktoken
chainlit
graphrag
marker-pdf
torch
setup_all.bat
